# frontend/types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Project Desert API",
    "description": "API for managing ascetical practices and spiritual growth",
    "version": "2.0.0"
  },
  "paths": {
...
    "/daily-readings/readings/{date}": {
      "get": {
        "tags": ["daily-readings"],
        "summary": "Get Mass Readings",
        "description": "Get Mass readings for a specific date. Checks database cache first,\nthen fetches from Universalis API if not cached.\nDate should be in YYYYMMDD format (e.g., 20260105).",
        "operationId": "get_mass_readings_daily_readings_readings__date__get",
        "parameters": [
          {
            "name": "date",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Date" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/daily-readings/notes": {
      "post": {
        "tags": ["daily-readings"],
        "summary": "Create Or Update Note",
        "description": "Create or update a daily reading note for a user on a specific date.",
        "operationId": "create_or_update_note_daily_readings_notes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DailyReadingNoteCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyReadingNoteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/daily-readings/notes/{user_id}/{date}": {
      "get": {
        "tags": ["daily-readings"],
        "summary": "Get Note By Date",
        "description": "Get a user's daily reading note for a specific date.\nReturns 404 if no note exists for that date.",
        "operationId": "get_note_by_date_daily_readings_notes__user_id___date__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "User Id" }
          },
          {
            "name": "date",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Date" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyReadingNoteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/daily-readings/notes/{user_id}": {
      "get": {
        "tags": ["daily-readings"],
        "summary": "Get All User Notes",
        "description": "Get all daily reading notes for a user, ordered by date descending.",
        "operationId": "get_all_user_notes_daily_readings_notes__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "User Id" }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "integer" }, { "type": "null" }],
              "default": 30,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DailyReadingNoteResponse"
                  },
                  "title": "Response Get All User Notes Daily Readings Notes  User Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/daily-readings/notes/{note_id}": {
      "delete": {
        "tags": ["daily-readings"],
        "summary": "Delete Note",
        "description": "Delete a daily reading note by ID.",
        "operationId": "delete_note_daily_readings_notes__note_id__delete",
        "parameters": [
          {
            "name": "note_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Note Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DailyReadingNoteCreate": {
        "properties": {
          "userId": { "type": "integer", "title": "Userid" },
          "date": { "type": "string", "title": "Date" },
          "notes": { "type": "string", "title": "Notes" }
        },
        "type": "object",
        "required": ["userId", "date", "notes"],
        "title": "DailyReadingNoteCreate",
        "description": "Request to create or update a daily reading note."
      },
      "DailyReadingNoteResponse": {
        "properties": {
          "id": { "type": "integer", "title": "Id" },
          "userId": { "type": "integer", "title": "Userid" },
          "date": { "type": "string", "title": "Date" },
          "notes": { "type": "string", "title": "Notes" },
          "createdAt": { "type": "string", "title": "Createdat" },
          "updatedAt": { "type": "string", "title": "Updatedat" }
        },
        "type": "object",
        "required": ["id", "userId", "date", "notes", "createdAt", "updatedAt"],
        "title": "DailyReadingNoteResponse",
        "description": "Daily reading note response."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    }
  }
}